% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acceptable_header}
\alias{acceptable_header}
\title{Get lookup table of acceptable header variables}
\usage{
acceptable_header()
}
\description{
Get lookup table of acceptable header variables
}
\examples{
acceptable_header()

}
