% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acceptable_nonunits}
\alias{acceptable_nonunits}
\title{Get lookup table of acceptable variable types for classes other than "units"}
\usage{
acceptable_nonunits()
}
\description{
Get lookup table of acceptable variable types for classes other than "units"
}
\examples{
acceptable_nonunits()

}
