% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acceptable_nonunits_6400}
\alias{acceptable_nonunits_6400}
\title{Get lookup table of acceptable nonunits for data}
\usage{
acceptable_nonunits_6400()
}
\description{
Get lookup table of acceptable nonunits for data
}
\examples{
acceptable_nonunits_6400()

}
