% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acceptable_types_6400}
\alias{acceptable_types_6400}
\title{Get vector of acceptable data types for licor 6400 files}
\usage{
acceptable_types_6400()
}
\description{
Get vector of acceptable data types for licor 6400 files
}
\examples{
acceptable_types_6400()

}
