% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acceptable_units_6400}
\alias{acceptable_units_6400}
\title{Get lookup table of acceptable units for data}
\usage{
acceptable_units_6400(replace = NULL)
}
\arguments{
\item{replace}{A named vector to replace the units for a 6400 file}
}
\description{
Get lookup table of acceptable units for data
}
\examples{
acceptable_units_6400()

}
