% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{apply_li6800_units}
\alias{apply_li6800_units}
\title{Applies the units for each variable exported from the LI-6800}
\usage{
apply_li6800_units(.x, unit_vector)
}
\arguments{
\item{.x}{data.frame. Raw data without units.}

\item{unit_vector}{Character. A vector of character values that can be
converted to valid scientific units. The vector should be in the same order
as variables in the data.}
}
\value{
The data.frame .x with units applied via unit_vector.
}
\description{
Applies the units for each variable exported from the LI-6800
}
