% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{combine_licor}
\alias{combine_licor}
\title{Combines any number of licor files, in the order they are passed.}
\usage{
combine_licor(combo)
}
\arguments{
\item{combo}{A list of licor objects, to be combined}
}
\value{
Returns a licor composed of all the passed licor objects
}
\description{
Combines any number of licor files, in the order they are passed.
}
\examples{
\dontrun{
 #nothing yet
}
}
