% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fileprep}
\alias{fileprep}
\title{A preparation function to specially format files in windows
For use only in windows machines, automatically called when reading in
LI-6800 files.}
\usage{
fileprep(file)
}
\arguments{
\item{file}{The name of the licor raw file to read.}
}
\value{
Nothing is returned, but the file is modified.
}
\description{
A preparation function to specially format files in windows
For use only in windows machines, automatically called when reading in
LI-6800 files.
}
