% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarks.R
\name{get_li6800_remarks}
\alias{get_li6800_remarks}
\title{Gets the remarks from an LI-6800 file}
\usage{
get_li6800_remarks(raw_lines)
}
\arguments{
\item{raw_lines}{Character. Vector of raw lines from LI-6800 file.}
}
\value{
A tibble of remarks with three columns named "row", "time", and "remark".
}
\description{
Gets the remarks from an LI-6800 file
}
