% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{get_li6800_types_line}
\alias{get_li6800_types_line}
\title{Gets the lines with types for each variable exported from the LI-6800}
\usage{
get_li6800_types_line(raw_lines)
}
\arguments{
\item{raw_lines}{Character. Vector of raw lines from LI-6800 file.}
}
\value{
An integer indexing which line in raw_lines contains the variable
types.
}
\description{
Gets the lines with types for each variable exported from the LI-6800
}
