% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_li6800_units}
\alias{get_li6800_units}
\title{Gets the units for each variable exported from the LI-6800}
\usage{
get_li6800_units(raw_lines, types_line, which)
}
\arguments{
\item{raw_lines}{Character. Vector of raw lines from LI-6800 file.}

\item{types_line}{Integer. Index of which line in raw_lines contains the variable
types.}

\item{which}{Integer. Vector of variable column indexes.}
}
\value{
A character vector of types for each variable. The index attribute
is the column number.
}
\description{
Gets the units for each variable exported from the LI-6800
}
