% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{get_remarks}
\alias{get_remarks}
\title{Outputs options for subsetting}
\usage{
get_remarks(x)
}
\arguments{
\item{x}{A licor object}
}
\value{
Returns the remarks present in the original file, if any.
}
\description{
Outputs options for subsetting
}
\examples{
\donttest{

}
}
