% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarks.R
\name{is_remark_6400}
\alias{is_remark_6400}
\title{is_remark_6400}
\usage{
is_remark_6400(.s)
}
\arguments{
\item{.s}{A string to be checked for remarks

Detect remarks in LI-6400 raw data}
}
\description{
is_remark_6400
}
