% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{licor}
\alias{licor}
\title{Helper for class licor. Creates licor class objects.}
\usage{
licor(file, dec = ".", replace = NULL)
}
\arguments{
\item{file}{The name of the licor raw file to read.}

\item{dec}{Whether the decimal is "." or ","}

\item{replace}{A named vector to replace the units for a 6400 file}
}
\value{
Returns an object of class licor, checked for correctness. Supports
6800 files currently. Correct licor objects have units on their variables
and an attribute containg the data from the header of the file.
}
\description{
Helper for class licor. Creates licor class objects.
}
\examples{

\dontrun{
licor("/path/filename")
}
}
