% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{licor_subset}
\alias{licor_subset}
\title{Creating a subset of the data}
\usage{
licor_subset(x, vec)
}
\arguments{
\item{x}{A licor object}

\item{vec}{A vector of subset options, and can be any permutation of the
options presented by the 'subset_options' function, for example, vec could
be 'c("Sys","GasEx")'}
}
\value{
Returns a data.frame with the subset of data specified
}
\description{
Creating a subset of the data
}
\examples{
\donttest{

}
}
