% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorer.R
\docType{package}
\name{licorer}
\alias{licorer}
\title{\code{licorer} package}
\description{
Read and process LI-COR photosynthesis data files in R
}
\details{
See the README on
\href{https://github.com/muir-lab/licorer}{GitHub}
}
