% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{new_licor}
\alias{new_licor}
\title{Constructor for class licor.}
\usage{
new_licor(file, dec = ".", replace = NULL)
}
\arguments{
\item{file}{The name of the file to read from}

\item{dec}{Whether to set the decimal as a "." or a ","}

\item{replace}{A named vector to replace the units for a 6400 file}
}
\value{
Object of class licor
}
\description{
Constructor for class licor.
}
\examples{
new_licor(system.file("extdata", "2019-05-06-0740_trillium_ovatum",
                      package = "licorer", mustWork = TRUE))
\dontrun{
new_licor("/path/filename")
}

}
