% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_li6400}
\alias{read_li6400}
\title{Reads data from LI-6400 file and organizes it into class licor.}
\usage{
read_li6400(file, dec = ".", replace = NULL)
}
\arguments{
\item{file}{The name of the file to read from}

\item{dec}{Whether to set the decimal as a "." or a ","}

\item{replace}{A named vector to replace the units for a 6400 file}
}
\value{
Returns a dataframe from raw LiCor 6400 files.
}
\description{
Reads data from LI-6400 file and organizes it into class licor.
}
\examples{
\dontrun{
read_li6800_raw(system.file("extdata", "2019-05-06-0740_trillium_ovatum",
                            package = "licorer", mustWork = TRUE))
read_li6800_raw("/path/filename")
}

}
