% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_li6400_raw}
\alias{read_li6400_raw}
\title{Reads data from raw LI-6400 files and organizes into class licor.}
\usage{
read_li6400_raw(file, dec = ".", replace = NULL)
}
\arguments{
\item{file}{The name of the file to read from}

\item{dec}{Whether to set the decimal as a "." or a ","}

\item{replace}{A named vector to replace the units for a 6400 file}
}
\value{
Returns an object of class licor from raw LI-6400 files.
}
\description{
Reads data from raw LI-6400 files and organizes into class licor.
}
