% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_li6800}
\alias{read_li6800}
\title{Reads data from LI-6800 file and organizes it into class licor.}
\usage{
read_li6800(file, decimal = ".")
}
\arguments{
\item{file}{A path to a file exported from the LI-6800.}

\item{decimal}{Whether to set the decimal as a "." or a ",".}
}
\value{
Returns a data.frame from raw LI-6800 file.
}
\description{
Reads data from LI-6800 file and organizes it into class licor.
}
\examples{
\dontrun{
read_li6800_raw(system.file("extdata", "2020-06-23_ECT_hib",
                            package = "licorer", mustWork = TRUE))
read_li6800_raw("/path/filename")
}

}
