% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_li6800_excel}
\alias{read_li6800_excel}
\title{Reads data from Excel LI-6800 files and organizes into class licor.}
\usage{
read_li6800_excel(file, decimal = ".")
}
\arguments{
\item{file}{A path to a file exported from the LI-6800.}

\item{decimal}{Whether to set the decimal as a "." or a ",".}
}
\description{
Reads data from Excel LI-6800 files and organizes into class licor.
}
