% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_li6800_raw}
\alias{read_li6800_raw}
\title{Reads data from raw LI-6800 files and organizes into class licor.}
\usage{
read_li6800_raw(file, decimal = ".")
}
\arguments{
\item{file}{A path to a file exported from the LI-6800.}

\item{decimal}{Whether to set the decimal as a "." or a ",".}
}
\value{
Returns an object of class 'licor' from raw LI-6800 files. Current
support for LI-6800 files only.
}
\description{
Reads data from raw LI-6800 files and organizes into class licor.
}
