% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{subset}
\alias{subset}
\title{Creating a subset of the data}
\usage{
subset(x, vec)
}
\arguments{
\item{x}{A licor object}

\item{vec}{A vector of subset options}
}
\value{
Returns a data.frame with the subset of data specified
}
\description{
Creating a subset of the data
}
\examples{
\donttest{

}
}
