% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{subset_logic}
\alias{subset_logic}
\title{Creates a boolean logical vector}
\usage{
subset_logic(x, vec)
}
\arguments{
\item{x}{A licor object}

\item{vec}{A vector of subset options}
}
\value{
Returns a logical vector in order to subset/perform operations
on data without being destructive or duplicating unnecissarily.
}
\description{
Creates a boolean logical vector
}
\examples{
\donttest{

}
}
