% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{subset_options}
\alias{subset_options}
\title{Outputs options for subsetting}
\usage{
subset_options(x)
}
\arguments{
\item{x}{A licor object}
}
\value{
Returns a vector of strings that can be used to subset the data
using the subset function.
}
\description{
Outputs options for subsetting
}
\examples{
\dontrun{
   subset_options()
}
}
