% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{validate_licor}
\alias{validate_licor}
\title{Validator for class licor.}
\usage{
validate_licor(x)
}
\arguments{
\item{x}{x}
}
\value{
x
}
\description{
Validator for class licor.
}
\examples{
\dontrun{
validate_licor(licor_object)
}
}
