% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acceptable_units}
\alias{acceptable_units}
\title{Get lookup table of acceptable units for data}
\usage{
acceptable_units()
}
\description{
Get lookup table of acceptable units for data
}
\examples{
acceptable_types()

}
