% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{fileprep}
\alias{fileprep}
\title{A preparation funtion to specially format files in windows
For use only in windows macines, automatically called when reading in licor
files.}
\usage{
fileprep(file)
}
\arguments{
\item{file}{The name of the licor raw file to read.}
}
\value{
Nothing is returned, but the file is modified.
}
\description{
A preparation funtion to specially format files in windows
For use only in windows macines, automatically called when reading in licor
files.
}
