% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{licor}
\alias{licor}
\title{Helper for class licor. Creates licor class objects.}
\usage{
licor(file, deci = ".")
}
\arguments{
\item{file}{The name of the licor raw file to read.}

\item{deci}{Wheter the decimal is "." or ","}
}
\value{
Returns an object of class licor, checked for correctness. Supports
6800 files currently. Correct licor objects have units on their variables
and an attribute containg the data from the header of the file.
}
\description{
Helper for class licor. Creates licor class objects.
}
\examples{
licor(system.file("extdata", "2019-05-06-0740_trillium_ovatum", package = "licorer", mustWork = TRUE))
\donttest{
licor("/path/filename")
}
}
