% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{new_licor}
\alias{new_licor}
\title{Constructor for class licor.}
\usage{
new_licor(filename, deci = ".")
}
\arguments{
\item{filename}{The name of the file to read from}

\item{deci}{Whether to set the decimal as a "." or a ","}
}
\value{
Object of class licor
}
\description{
Constructor for class licor.
}
\examples{
new_licor(system.file("extdata", "2019-05-06-0740_trillium_ovatum", package = "licorer", mustWork = TRUE))
new_licor("/path/filename")

}
