% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_li6800}
\alias{read_li6800}
\title{Reads data from LI-6800 file and organizes it into class licor.}
\usage{
read_li6800(file, dec = ".")
}
\arguments{
\item{file}{The name of the file to read from}

\item{dec}{Whether to set the decimal as a "." or a ","}
}
\value{
Returns a dataframe from raw LiCor 6800 files.
}
\description{
Reads data from LI-6800 file and organizes it into class licor.
}
\examples{
\dontrun{
read_li6800_raw(system.file("extdata", "2020-06-23_ECT_hib",
                            package = "licorer", mustWork = TRUE))
read_li6800_raw("/path/filename")
}

}
