% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_li6800}
\alias{read_li6800}
\alias{read_li6800_excel}
\alias{read_li6800_raw}
\title{Reads data from LI-6800 file and organizes it into class licor.}
\usage{
read_li6800(file, dec = ".")

read_li6800_excel(file, dec = ".")

read_li6800_raw(file, dec = ".")
}
\value{
Returns a dataframe from raw LiCor files. Current support
for LiCor 6800 files only. LiCor 6400 file reading will be supported
in a later version.

Returns an object of class licor from raw LI-6800 files. Current
support for LI-6800 files only. LI-6400 file reading will be supported
in a later version.
}
\description{
Reads data from LI-6800 file and organizes it into class licor.

Reads data from Excel LI-6800 files and organizes into class licor.

Reads data from raw LI-6800 files and organizes into class licor.
}
\examples{
\donttest{
read_li6800_raw(system.file("extdata", "2019-05-06-0740_trillium_ovatum", package = "licorer", mustWork = TRUE))
read_li6800_raw("/path/filename")
}

}
