% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_li6800_excel}
\alias{read_li6800_excel}
\title{Reads data from Excel LI-6800 files and organizes into class licor.}
\usage{
read_li6800_excel(file, dec = ".")
}
\arguments{
\item{file}{The name of the file to read from}

\item{dec}{Whether to set the decimal as a "." or a ","}
}
\description{
Reads data from Excel LI-6800 files and organizes into class licor.
}
