% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_li6800_raw}
\alias{read_li6800_raw}
\title{Reads Data from raw LiCor files and organizes into class licor.}
\usage{
read_li6800_raw(filename, deci = ".")
}
\arguments{
\item{filename}{The name of the file to read from}

\item{deci}{Whether to set the decimal as a "." or a ","}
}
\value{
Returns an object of class licor from raw LiCor files. Current support
for LiCor 6800 files only. LiCor 6400 file reading will be supported
in a later version.
}
\description{
Reads Data from raw LiCor files and organizes into class licor.
}
\examples{
\donttest{
read_li6800_raw(system.file("extdata", "2019-05-06-0740_trillium_ovatum", package = "licorer", mustWork = TRUE))
read_li6800_raw("/path/filename")
}
}
