% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_li6800_raw}
\alias{read_li6800_raw}
\title{Reads data from raw LI-6800 files and organizes into class licor.}
\usage{
read_li6800_raw(file, dec = ".")
}
\arguments{
\item{file}{The name of the file to read from}

\item{dec}{Whether to set the decimal as a "." or a ","}
}
\value{
Returns an object of class licor from raw LI-6800 files. Current
support for LI-6800 files only. LI-6400 file reading will be supported
in a later version.
}
\description{
Reads data from raw LI-6800 files and organizes into class licor.
}
